+++ forui/test/src/widgets/tile/tile_group_merge_golden_test.dart
@@ -25,10 +25,10 @@ void main() {
                 label: const Text('Network'),
                 children: [
                   FTile(
-                    prefixIcon: const Icon(FIcons.wifi),
+                    prefix: const Icon(FIcons.wifi),
                     title: const Text('WiFi'),
                     details: const Text('Forus Labs (5G)'),
-                    suffixIcon: const Icon(FIcons.chevronRight),
+                    suffix: const Icon(FIcons.chevronRight),
                     onPress: () {},
                   ),
                 ],
@@ -49,18 +49,18 @@ void main() {
             label: const Text('Network'),
             children: [
               FTile(
-                prefixIcon: const Icon(FIcons.wifi),
+                prefix: const Icon(FIcons.wifi),
                 title: const Text('WiFi'),
                 details: const Text('Forus Labs (5G)'),
-                suffixIcon: const Icon(FIcons.chevronRight),
+                suffix: const Icon(FIcons.chevronRight),
                 onPress: () {},
               ),
               FTile(
-                prefixIcon: const Icon(FIcons.bluetooth),
+                prefix: const Icon(FIcons.bluetooth),
                 title: const Text('Bluetooth'),
                 subtitle: const Text('Fee, Fo, Fum'),
                 details: const Text('Forus Labs (5G)'),
-                suffixIcon: const Icon(FIcons.chevronRight),
+                suffix: const Icon(FIcons.chevronRight),
                 onPress: () {},
               ),
             ],
@@ -86,18 +86,18 @@ void main() {
             children: [
               FTile(
                 enabled: false,
-                prefixIcon: const Icon(FIcons.wifi),
+                prefix: const Icon(FIcons.wifi),
                 title: const Text('WiFi'),
                 details: const Text('Forus Labs (5G)'),
-                suffixIcon: const Icon(FIcons.chevronRight),
+                suffix: const Icon(FIcons.chevronRight),
                 onPress: () {},
               ),
               FTile(
-                prefixIcon: const Icon(FIcons.bluetooth),
+                prefix: const Icon(FIcons.bluetooth),
                 title: const Text('Bluetooth'),
                 subtitle: const Text('Fee, Fo, Fum'),
                 details: const Text('Forus Labs (5G)'),
-                suffixIcon: const Icon(FIcons.chevronRight),
+                suffix: const Icon(FIcons.chevronRight),
                 onPress: () {},
               ),
             ],
@@ -123,22 +123,22 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.wifi),
+                      prefix: const Icon(FIcons.wifi),
                       title: const Text('WiFi'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.mail),
+                      prefix: const Icon(FIcons.mail),
                       title: const Text('Mail'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.bluetooth),
+                      prefix: const Icon(FIcons.bluetooth),
                       title: const Text('Bluetooth'),
                       subtitle: const Text('Fee, Fo'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -146,16 +146,16 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.wifi),
+                      prefix: const Icon(FIcons.wifi),
                       title: const Text('WiFi'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.bluetooth),
+                      prefix: const Icon(FIcons.bluetooth),
                       title: const Text('Bluetooth'),
                       subtitle: const Text('Fee, Fo'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -163,16 +163,16 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.wifi),
+                      prefix: const Icon(FIcons.wifi),
                       title: const Text('WiFi'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.bluetooth),
+                      prefix: const Icon(FIcons.bluetooth),
                       title: const Text('Bluetooth'),
                       subtitle: const Text('Fee, Fo'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -198,9 +198,9 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.wifi),
+                      prefix: const Icon(FIcons.wifi),
                       title: const Text('WiFi'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -208,14 +208,14 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.mail),
+                      prefix: const Icon(FIcons.mail),
                       title: const Text('Mail'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.bluetooth),
+                      prefix: const Icon(FIcons.bluetooth),
                       title: const Text('Bluetooth'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -244,9 +244,9 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.wifi),
+                      prefix: const Icon(FIcons.wifi),
                       title: const Text('WiFi'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -254,14 +254,14 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.mail),
+                      prefix: const Icon(FIcons.mail),
                       title: const Text('Mail'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.bluetooth),
+                      prefix: const Icon(FIcons.bluetooth),
                       title: const Text('Bluetooth'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -292,16 +292,16 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.wifi),
+                      prefix: const Icon(FIcons.wifi),
                       title: const Text('WiFi'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                     FTile(
                       focusNode: focusNode,
-                      prefixIcon: const Icon(FIcons.mail),
+                      prefix: const Icon(FIcons.mail),
                       title: const Text('Mail'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                   ],
@@ -309,15 +309,15 @@ void main() {
                 FTileGroup(
                   children: [
                     FTile(
-                      prefixIcon: const Icon(FIcons.bluetooth),
+                      prefix: const Icon(FIcons.bluetooth),
                       title: const Text('Bluetooth'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                       onPress: () {},
                     ),
                     FTile(
-                      prefixIcon: const Icon(FIcons.arrowDown),
+                      prefix: const Icon(FIcons.arrowDown),
                       title: const Text('Last'),
-                      suffixIcon: const Icon(FIcons.chevronRight),
+                      suffix: const Icon(FIcons.chevronRight),
                     ),
                   ],
                 ),
@@ -351,9 +351,9 @@ void main() {
               FTileGroup(
                 children: [
                   FTile(
-                    prefixIcon: const Icon(FIcons.wifi),
+                    prefix: const Icon(FIcons.wifi),
                     title: const Text('WiFi'),
-                    suffixIcon: const Icon(FIcons.chevronRight),
+                    suffix: const Icon(FIcons.chevronRight),
                     onPress: () {},
                   ),
                 ],
@@ -361,9 +361,9 @@ void main() {
               FTileGroup(
                 children: [
                   FTile(
-                    prefixIcon: const Icon(FIcons.wifi),
+                    prefix: const Icon(FIcons.wifi),
                     title: const Text('WiFi'),
-                    suffixIcon: const Icon(FIcons.chevronRight),
+                    suffix: const Icon(FIcons.chevronRight),
                     onPress: () {},
                   ),
                 ],
@@ -388,9 +388,9 @@ void main() {
               FTileGroup(
                 children: [
                   FTile(
-                    prefixIcon: const Icon(FIcons.wifi),
+                    prefix: const Icon(FIcons.wifi),
                     title: const Text('WiFi'),
-                    suffixIcon: const Icon(FIcons.chevronRight),
+                    suffix: const Icon(FIcons.chevronRight),
                     onPress: () {},
                   ),
                 ],
@@ -398,9 +398,9 @@ void main() {
               FTileGroup(
                 children: [
                   FTile(
-                    prefixIcon: const Icon(FIcons.wifi),
+                    prefix: const Icon(FIcons.wifi),
                     title: const Text('WiFi'),
-                    suffixIcon: const Icon(FIcons.chevronRight),
+                    suffix: const Icon(FIcons.chevronRight),
                     onPress: () {},
                   ),
                 ],
@@ -421,10 +421,10 @@ void main() {
               FTileGroup(
                 children: [
                   FTile(
-                    prefixIcon: const Icon(FIcons.wifi),
+                    prefix: const Icon(FIcons.wifi),
                     title: const Text('WiFi'),
                     details: const Text('FL (5G)'),
-                    suffixIcon: const Icon(FIcons.chevronRight),
+                    suffix: const Icon(FIcons.chevronRight),
                     onPress: () {},
                   ),
                 ],
@@ -458,15 +458,15 @@ void main() {
               divider: FTileDivider.none,
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.list, color: Colors.transparent),
+                  prefix: const Icon(FIcons.list, color: Colors.transparent),
                   title: const Text('Personalization'),
-                  suffixIcon: const Icon(FIcons.user),
+                  suffix: const Icon(FIcons.user),
                   onPress: () {},
                 ),
                 FTile(
-                  prefixIcon: const Icon(FIcons.list, color: Colors.transparent),
+                  prefix: const Icon(FIcons.list, color: Colors.transparent),
                   title: const Text('Network'),
-                  suffixIcon: const Icon(FIcons.appWindowMac),
+                  suffix: const Icon(FIcons.appWindowMac),
                   onPress: () {},
                 ),
               ],
@@ -475,15 +475,15 @@ void main() {
               divider: FTileDivider.none,
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.list, color: Colors.transparent),
+                  prefix: const Icon(FIcons.list, color: Colors.transparent),
                   title: const Text('List View'),
-                  suffixIcon: const Icon(FIcons.list),
+                  suffix: const Icon(FIcons.list),
                   onPress: () {},
                 ),
                 FTile(
-                  prefixIcon: const Icon(FIcons.list, color: Colors.transparent),
+                  prefix: const Icon(FIcons.list, color: Colors.transparent),
                   title: const Text('Grid View'),
-                  suffixIcon: const Icon(FIcons.layoutGrid),
+                  suffix: const Icon(FIcons.layoutGrid),
                   onPress: () {},
                 ),
               ],
@@ -508,10 +508,10 @@ void main() {
               error: const Text('This should not appear'),
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
               ],
@@ -536,10 +536,10 @@ void main() {
               error: const Text('This should not appear'),
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
               ],
@@ -550,10 +550,10 @@ void main() {
               error: const Text('This should not appear'),
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
               ],
@@ -575,29 +575,29 @@ void main() {
               style: FThemes.green.dark.tileGroupStyle,
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
                 FTile(
                   style: FThemes.blue.dark.tileGroupStyle.untappableTileStyle,
-                  prefixIcon: const Icon(FIcons.bluetooth),
+                  prefix: const Icon(FIcons.bluetooth),
                   title: const Text('Bluetooth'),
                   subtitle: const Text('Fee, Fo'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                 ),
               ],
             ),
             FTileGroup(
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
               ],
@@ -619,18 +619,18 @@ void main() {
             FTileGroup(
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
                 FTile(
-                  prefixIcon: const Icon(FIcons.bluetooth),
+                  prefix: const Icon(FIcons.bluetooth),
                   title: const Text('Bluetooth'),
                   subtitle: const Text('Fee, Fo'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                 ),
               ],
             ),
@@ -638,10 +638,10 @@ void main() {
               enabled: true,
               children: [
                 FTile(
-                  prefixIcon: const Icon(FIcons.wifi),
+                  prefix: const Icon(FIcons.wifi),
                   title: const Text('WiFi'),
                   details: const Text('FL (5G)'),
-                  suffixIcon: const Icon(FIcons.chevronRight),
+                  suffix: const Icon(FIcons.chevronRight),
                   onPress: () {},
                 ),
               ],